+++ module/constants/activeEffectsTypes.js
@@ -1,4 +1,4 @@
-export default  {
+export default {
   spellcasting: {
     category: "magic",
     type: "spellcasting",
@@ -6,12 +6,11 @@ export default  {
     key: "data.bonuses.arts.spellcasting"
   },
   penetration: {
-  //   label: "arm5e.sheet.effects.type.spellcasting"
-  //   key: "data.bonuses.arts.laboratory"
-  // }
-    type: 'penetration',
-    key: 'data.bonuses.arts.penetration'
-  }
+    category: "magic",
+    type: "penetration",
+    label: "arm5e.skill.penetration",
+    key: "data.bonuses.arts.penetration"
+  },
   cr: {
     category: "magic",
     type: "cr",
@@ -107,6 +106,7 @@ export default  {
     type: "size",
     label: "arm5e.sheet.size",
     key: "data.vitals.siz.value"
+  }
 
   // soak: {
   //   category: "traits",
@@ -120,4 +120,4 @@ export default  {
   //   label: "arm5e.sheet.strength",
   //   key: "data.characteristics.str.value"
   // }
-}
+};